Backend/AnsibleAutomation/requirements.txt
flask==2.3.0
flask-cors==4.0.0
groq==0.10.0
pymongo==4.6.0
python-dotenv==1.0.0
